pyproject.toml
[project]
name = "dmarc-report-manager"
version = "0.2.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "dnspython>=2.8.0",
    "fastapi>=0.128.0",
    "pydantic>=2.12.5",
    "python-multipart>=0.0.22",
    "uvicorn>=0.40.0",
    "xmltodict>=1.0.2",
    "pytest>=8.0.0",
    "httpx>=0.27.0",
    "passlib[bcrypt]>=1.7.4",
    "pyjwt>=2.10.0",
    "python-dotenv>=1.2.1",
    "slowapi>=0.1.9",
    "fpdf2>=2.8.5",
]

